move module parsing to separate file
